backend/flaskapi/requirements.txt
flask
requests
chromadb
sentence-transformers
python-dotenv
langdetect
googletrans
